+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2018-11-13
 
+- mission-new-button: use executorId passed from Missions page
 - adjustment in apollo.js to indicate `uri` for regular
   dockerized release, and for local development
 - docker related adjustments while deploying on tsauv